Direction fix, new dump folder structure

+++ CTFAK/MMFParser/EXE/Loaders/Events/Parameters/Position.cs
@@ -5,12 +5,12 @@ namespace CTFAK.MMFParser.EXE.Loaders.Events.Parameters
     public class Position : ParameterCommon
     {
         public int ObjectInfoParent;
-        public ushort Flags;
+        public short Flags;
         public int X;
         public int Y;
         public int Slope;
         public int Angle;
-        public float Direction;
+        public int Direction;
         public int TypeParent;
         public int ObjectInfoList;
         public int Layer;            
@@ -19,12 +19,12 @@ namespace CTFAK.MMFParser.EXE.Loaders.Events.Parameters
         public override void Read()
         {
             ObjectInfoParent = Reader.ReadInt16();
-            Flags = Reader.ReadUInt16();
+            Flags = Reader.ReadInt16();
             X = Reader.ReadInt16();
             Y = Reader.ReadInt16();
             Slope = Reader.ReadInt16();
             Angle = Reader.ReadInt16();
-            Direction = Reader.ReadSingle();
+            Direction = Reader.ReadInt32();
             TypeParent = Reader.ReadInt16();
             ObjectInfoList = Reader.ReadInt16();
             Layer = Reader.ReadInt16();
@@ -33,12 +33,12 @@ namespace CTFAK.MMFParser.EXE.Loaders.Events.Parameters
         public override void Write(ByteWriter Writer)
         {
             Writer.WriteInt16((short) ObjectInfoParent);
-            Writer.WriteUInt16(Flags);
+            Writer.WriteInt16(Flags);
             Writer.WriteInt16((short) X);
             Writer.WriteInt16((short) Y);
             Writer.WriteInt16((short) Slope);
             Writer.WriteInt16((short) Angle);
-            Writer.WriteSingle(Direction);
+            Writer.WriteInt32(Direction);
             Writer.WriteInt16((short) TypeParent);
             Writer.WriteInt16((short) ObjectInfoList);
             Writer.WriteInt16((short) Layer);
